Properly fix up rootdir/android.mk
If defining all GSM and ending the ifeq then you
If m7spr foo else bar, and else everything for version --> this results in the GSM defing defined again in the else.

You can ifeq, else if sprint, else (everything else, which is VZW)
Or perhaps the best way is to define everything.  Which is what I did, so no more duped definitions

Change-Id: I3db0af8ad78ba44a271c4c647024c9f21caff04a